Refactor and clean model architectures
